+++ src/pygaps/parsing/csvinterface.py
@@ -1,12 +1,15 @@
 """The csv parsing interface."""
 
 import ast
+from io import StringIO
 
 import pandas
 
+from ..core.modelisotherm import Isotherm
 from ..core.modelisotherm import ModelIsotherm
 from ..core.pointisotherm import PointIsotherm
 from ..modelling import get_isotherm_model
+from ..utilities.exceptions import ParsingError
 
 
 def _is_float(s):
@@ -57,72 +60,87 @@ def _to_string(s):
     return str(s)
 
 
-def isotherm_to_csv(isotherm, path, separator=','):
+def isotherm_to_csv(isotherm, path=None, separator=','):
     """
-    Convert isotherm into a CSV file.
+    Convert isotherm into a CSV representation.
+
+    If the path is specified, the isotherm is saved as a file,
+    otherwise it is returned as a string.
 
     Parameters
     ----------
     isotherm : Isotherm
         Isotherm to be written to csv.
-    path : str
+    path : str, None
         Path to the file to be written.
     separator : str, optional
         Separator used int the csv file. Defaults to '',''.
 
     """
-    with open(path, mode='w', newline='\n') as file:
+    output = StringIO()
 
-        isotherm_data = isotherm.to_dict()
+    isotherm_data = isotherm.to_dict()
 
-        file.writelines([
-            x + separator + _to_string(y) + '\n'
-            for (x, y) in isotherm_data.items()
+    output.writelines([
+        x + separator + _to_string(y) + '\n'
+        for (x, y) in isotherm_data.items()
+    ])
+
+    if isinstance(isotherm, PointIsotherm):
+
+        # get headings in an ordered way
+        headings = [
+            isotherm.pressure_key,
+            isotherm.loading_key,
+        ]
+        if isotherm.other_keys:
+            headings.extend(isotherm.other_keys)
+
+        # also get the branch data in a regular format
+        headings.append('branch')
+        data = isotherm.data_raw[headings]
+        data['branch'] = data['branch'].replace(False,
+                                                'ads').replace(True, 'des')
+
+        output.write('data:[pressure,loading,[otherdata],branch data]\n')
+        data.to_csv(output, sep=separator, index=False, header=True)
+
+    elif isinstance(isotherm, ModelIsotherm):
+
+        output.write('model:[name and parameters]\n')
+        output.write(('name' + separator + isotherm.model.name + '\n'))
+        output.write(
+            ('rmse' + separator + _to_string(isotherm.model.rmse) + '\n')
+        )
+        output.write((
+            'pressure range' + separator +
+            _to_string(isotherm.model.pressure_range) + '\n'
+        ))
+        output.write((
+            'loading range' + separator +
+            _to_string(isotherm.model.loading_range) + '\n'
+        ))
+        output.writelines([
+            param + separator + str(isotherm.model.params[param]) + '\n'
+            for param in isotherm.model.params
         ])
 
-        if isinstance(isotherm, PointIsotherm):
-
-            # get headings in an ordered way
-            headings = [
-                isotherm.pressure_key,
-                isotherm.loading_key,
-            ]
-            if isotherm.other_keys:
-                headings.extend(isotherm.other_keys)
-
-            # also get the branch data in a regular format
-            headings.append('branch')
-            data = isotherm.data_raw[headings]
-            data['branch'] = data['branch'].replace(False, 'ads').replace(
-                True, 'des')
-
-            file.write('data:[pressure,loading,[otherdata],branch data]\n')
-            data.to_csv(file, sep=separator, index=False, header=True)
-
-        elif isinstance(isotherm, ModelIsotherm):
-
-            file.write('model:[name and parameters]\n')
-            file.write(('name' + separator + isotherm.model.name + '\n'))
-            file.write(
-                ('rmse' + separator + _to_string(isotherm.model.rmse) + '\n'))
-            file.write(('pressure range' + separator +
-                        _to_string(isotherm.model.pressure_range) + '\n'))
-            file.write(('loading range' + separator +
-                        _to_string(isotherm.model.loading_range) + '\n'))
-            file.writelines([
-                param + separator + str(isotherm.model.params[param]) + '\n'
-                for param in isotherm.model.params
-            ])
+    if path:
+        with open(path, mode='w', newline='\n') as file:
+            print(output.getvalue(), file=file)
+    else:
+        return output.getvalue()
 
 
-def isotherm_from_csv(path, separator=',', **isotherm_parameters):
+def isotherm_from_csv(str_or_path, separator=',', **isotherm_parameters):
     """
     Load an isotherm from a CSV file.
 
     Parameters
     ----------
-    path : str
-        Path to the file to be read.
+    str_or_path : str
+        The isotherm in a CSV string format or a path
+        to where one can be read.
     separator : str, optional
         Separator used int the csv file. Defaults to `,`.
     isotherm_parameters :
@@ -131,84 +149,101 @@ def isotherm_from_csv(path, separator=',', **isotherm_parameters):
     Returns
     -------
     Isotherm
-        The isotherm contained in the csv file.
+        The isotherm contained in the csv string or file.
 
     """
-    with open(path) as file:
-        line = file.readline().rstrip()
-        raw_dict = {}
+    try:
+        with open(str_or_path) as f:
+            raw_csv = StringIO(f.read())
+    except OSError:
+        try:
+            raw_csv = StringIO(str_or_path)
+        except Exception:
+            raise ParsingError(
+                "Could not parse JSON isotherm. "
+                "The `str_or_path` is invalid or does not exist. "
+            )
 
-        while not (line.startswith('data') or line.startswith('model')
-                   or line == ""):
+    line = raw_csv.readline().rstrip()
+    raw_dict = {}
+
+    while not (
+        line.startswith('data') or line.startswith('model') or line == ""
+    ):
+        values = line.split(sep=separator)
+
+        if _is_bool(values[1]):
+            val = _to_bool(values[1])
+        elif _is_float(values[1]):
+            val = float(values[1])
+        elif _is_list(values[1]):
+            val = _from_list(values[1])
+        else:
+            val = values[1]
+        raw_dict.update({values[0]: val})
+        line = raw_csv.readline().rstrip()
+
+    # Update dictionary with any user parameters
+    raw_dict.update(isotherm_parameters)
+
+    # Now read specific type of isotherm (Point, Model, Base)
+    if line.startswith('data'):
+        data = pandas.read_csv(raw_csv, sep=separator)
+
+        # process isotherm branches if they exist
+        raw_dict['branch'] = 'guess'
+        if 'branch' in data.columns:
+            raw_dict['branch'] = data['branch'].replace('ads', False).replace(
+                'des', True
+            ).values
+
+        # generate other keys
+        other_keys = [
+            column for column in data.columns.values
+            if column not in [data.columns[0], data.columns[1], 'branch']
+        ]
+
+        isotherm = PointIsotherm(
+            isotherm_data=data,
+            pressure_key=data.columns[0],
+            loading_key=data.columns[1],
+            other_keys=other_keys,
+            **raw_dict
+        )
+
+    elif line.startswith('model'):
+
+        model = {}
+        line = raw_csv.readline().rstrip()
+        model['name'] = line.split(sep=separator)[1]
+        line = raw_csv.readline().rstrip()
+        model['rmse'] = line.split(sep=separator)[1]
+        line = raw_csv.readline().rstrip()
+        model['pressure_range'] = _from_list(line.split(sep=separator)[1])
+        line = raw_csv.readline().rstrip()
+        model['loading_range'] = _from_list(line.split(sep=separator)[1])
+        line = raw_csv.readline().rstrip()
+        model['parameters'] = {}
+        while line != "":
             values = line.split(sep=separator)
+            model['parameters'][values[0]] = float(values[1])
+            line = raw_csv.readline().rstrip()
 
-            if _is_bool(values[1]):
-                val = _to_bool(values[1])
-            elif _is_float(values[1]):
-                val = float(values[1])
-            elif _is_list(values[1]):
-                val = _from_list(values[1])
-            else:
-                val = values[1]
-            raw_dict.update({values[0]: val})
-            line = file.readline().rstrip()
+        new_mod = get_isotherm_model(model['name'])
+        new_mod.rmse = model['rmse']
+        new_mod.pressure_range = model['pressure_range']
+        new_mod.loading_range = model['loading_range']
+        for param in new_mod.params:
+            try:
+                new_mod.params[param] = model['parameters'][param]
+            except KeyError as err:
+                raise KeyError(
+                    f"The CSV is missing parameter '{param}'"
+                ) from err
 
-        if line.startswith('data'):
-            data = pandas.read_csv(file, sep=separator)
+        isotherm = ModelIsotherm(model=new_mod, **raw_dict)
 
-            # process isotherm branches if they exist
-            raw_dict['branch'] = 'guess'
-            if 'branch' in data.columns:
-                raw_dict['branch'] = data['branch'].replace(
-                    'ads', False).replace('des', True).values
-
-            # generate other keys
-            other_keys = [
-                column for column in data.columns.values
-                if column not in [data.columns[0], data.columns[1], 'branch']
-            ]
-
-            # Update dictionary with any user parameters
-            raw_dict.update(isotherm_parameters)
-
-            isotherm = PointIsotherm(isotherm_data=data,
-                                     pressure_key=data.columns[0],
-                                     loading_key=data.columns[1],
-                                     other_keys=other_keys,
-                                     **raw_dict)
-
-        if line.startswith('model'):
-
-            model = {}
-            line = file.readline().rstrip()
-            model['name'] = line.split(sep=separator)[1]
-            line = file.readline().rstrip()
-            model['rmse'] = line.split(sep=separator)[1]
-            line = file.readline().rstrip()
-            model['pressure_range'] = _from_list(line.split(sep=separator)[1])
-            line = file.readline().rstrip()
-            model['loading_range'] = _from_list(line.split(sep=separator)[1])
-            line = file.readline().rstrip()
-            model['parameters'] = {}
-            while line != "":
-                values = line.split(sep=separator)
-                model['parameters'][values[0]] = float(values[1])
-                line = file.readline().rstrip()
-
-            new_mod = get_isotherm_model(model['name'])
-            new_mod.rmse = model['rmse']
-            new_mod.pressure_range = model['pressure_range']
-            new_mod.loading_range = model['loading_range']
-            for param in new_mod.params:
-                try:
-                    new_mod.params[param] = model['parameters'][param]
-                except KeyError as err:
-                    raise KeyError(
-                        f"The CSV is missing parameter '{param}'") from err
-
-            # Update dictionary with any user parameters
-            raw_dict.update(isotherm_parameters)
-
-            isotherm = ModelIsotherm(model=new_mod, **raw_dict)
+    else:
+        isotherm = Isotherm(**raw_dict)
 
     return isotherm